+++ toolbox/utils/reuse/dry-run.sh
@@ -1,0 +1,96 @@
+#!/bin/bash
+# Software Name: floss-toolbox
+# SPDX-FileCopyrightText: Copyright (c) Orange SA
+# SPDX-License-Identifier: Apache-2.0
+#
+# This software is distributed under the Apache 2.0 license,
+# the text of which is available at https://opensource.org/license/apache-2-0
+# or see the "LICENSE.txt" file for more details.
+#
+# Authors: See CONTRIBUTORS.txt
+# Software description: A toolbox of scripts to help work of forges admins and open source referents
+
+# Since...............: 03/04/2024
+# Description.........: Make a dry-run of the reuse features to check if everything is ready to use
+# Version.............: 1.0.0
+
+set -eu
+
+# Couts
+# -----
+
+NUMBER_OF_CHECKS=0
+NUMBER_OF_SUCCESS=0
+NUMBER_OF_WARNINGS=0
+NUMBER_OF_ERRORS=0
+
+# Utils
+# -----
+
+# $1 - File name to test
+CheckIfFileExists(){
+    if [ ! -f "$1" ]; then
+        echo "⛔  ERROR: The file '$1' does not exist"
+        NUMBER_OF_ERRORS=$((NUMBER_OF_ERRORS+1))
+    else
+        echo "✅  Cool! The file '$1' exists"
+        NUMBER_OF_SUCCESS=$((NUMBER_OF_SUCCESS+1))
+    fi
+    NUMBER_OF_CHECKS=$((NUMBER_OF_CHECKS+1))
+}
+
+# $1 - Runtime name
+# $2 - Command to check runtime
+# $3 - Expected / suggested version
+CheckIfRuntimeExists(){
+    runtime_version=`$2`
+    if [ $? != "0" ]; then # Exist status != 0 so runtime version check failed ; we assume the runtime is missing
+        echo "❌  WARNING: It seems '$1' is not ready"
+        NUMBER_OF_ERRORS=$((NUMBER_OF_ERRORS+1))        
+    else 
+        echo "✅  Cool! '$1' is available"
+        echo -e "\t🔎  You should check if the version is at least '$3'"
+        NUMBER_OF_SUCCESS=$((NUMBER_OF_SUCCESS+1))
+    fi
+    NUMBER_OF_CHECKS=$((NUMBER_OF_CHECKS+1))
+}
+
+# GitHub features
+# ---------------
+
+echo -e "\n---------------------------------"
+echo "Assertions for the text generator"
+echo "---------------------------------"
+
+echo -e "\nCheck files..."
+
+CheckIfFileExists "./README.md"
+CheckIfFileExists "./update-sources-header.py"
+CheckIfFileExists "./requirements.txt"
+CheckIfFileExists "./_/licenses.py"
+
+# Runtimes and tools
+# ------------------
+
+echo -e "\n-----------------------"
+echo "Assertions for runtimes"
+echo "-----------------------"
+
+echo -e "\nCheck for Python3..."
+CheckIfRuntimeExists "Python3" "python3 --version" "3.8.5"
+
+echo -e "\nCheck for reuse-tool..."
+CheckIfRuntimeExists "reuse-tool" "reuse --version" "3.0.1"
+
+# Conclusion
+# ----------
+
+echo -e "\n----------"
+echo "Conclusion"
+echo "----------"
+
+echo -e "\nDry-run done! See the logs above to check all points controls."
+echo -e "\tNumber of controls.......: $NUMBER_OF_CHECKS"
+echo -e "\tNumber of success........: $NUMBER_OF_SUCCESS"
+echo -e "\tNumber of warnings.......: $NUMBER_OF_WARNINGS"
+echo -e "\tNumber of errors.........: $NUMBER_OF_ERRORS"